pyproject.toml
[project]
name = "content-flux"
version = "0.1.0"
description = "Content Flux - AI-powered content creation and distribution pipeline using PortiaAI"
authors = [
    { name = "Content Flow Team", email = "team@contentflow.ai" }
]
readme = "README.md"
requires-python = ">=3.11,<3.14"

dependencies = [
    "python-dotenv>=1.1.1",
    "portia-sdk-python[all]==0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = []

[project.scripts]
content-flux = "main:main"
